+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.0 - 2022-12-28
+### Fixed
+- Robustly parse two letter status codes from `git status --porcelain`.
+    - Any non-blank non-? status code in column 1 now considered as "Staged" (Index).
+    - Any non-blank non-? status code in column 2 now considered as  "Modified" (Working Tree).
+
 ## 1.2.0 - 2022-12-28
 ### Changed
 - Include git rename operations as `.IsStaged`.